Upgrage crdVersion to v1 and add multigroup
