+++ k8s/Dockerfile
@@ -11,7 +11,7 @@ RUN go mod download
 
 # Copy the go source
 COPY main.go main.go
-COPY api/ api/
+COPY apis/ apis/
 COPY config/scheduler/config.yaml scheduler-config.yaml
 COPY controllers/ controllers/
 COPY generated/ generated/
