+++ k8s/PROJECT
@@ -1,36 +1,34 @@
 domain: v6d.io
 layout:
 - go.kubebuilder.io/v3
-plugins:
-  manifests.sdk.operatorframework.io/v2: {}
-  scorecard.sdk.operatorframework.io/v2: {}
+multigroup: true
 projectName: vineyard
 repo: github.com/v6d-io/v6d/k8s
 resources:
 - api:
-    crdVersion: v1beta1
+    crdVersion: v1
     namespaced: true
   controller: true
   domain: v6d.io
   group: k8s
   kind: LocalObject
-  path: github.com/v6d-io/v6d/k8s/api/k8s/v1alpha1
+  path: github.com/v6d-io/v6d/k8s/apis/k8s/v1alpha1
   version: v1alpha1
   webhooks:
     defaulting: true
     validation: true
-    webhookVersion: v1beta1
+    webhookVersion: v1
 - api:
-    crdVersion: v1beta1
+    crdVersion: v1
     namespaced: true
   controller: true
   domain: v6d.io
   group: k8s
   kind: GlobalObject
-  path: github.com/v6d-io/v6d/k8s/api/k8s/v1alpha1
+  path: github.com/v6d-io/v6d/k8s/apis/k8s/v1alpha1
   version: v1alpha1
   webhooks:
     defaulting: true
     validation: true
-    webhookVersion: v1beta1
+    webhookVersion: v1
 version: "3"
