+++ k8s/config/webhook/manifests.yaml
@@ -7,7 +7,6 @@ metadata:
 webhooks:
 - admissionReviewVersions:
   - v1
-  - v1beta1
   clientConfig:
     service:
       name: webhook-service
@@ -28,7 +27,6 @@ webhooks:
   sideEffects: NoneOnDryRun
 - admissionReviewVersions:
   - v1
-  - v1beta1
   clientConfig:
     service:
       name: webhook-service
@@ -56,7 +54,6 @@ metadata:
 webhooks:
 - admissionReviewVersions:
   - v1
-  - v1beta1
   clientConfig:
     service:
       name: webhook-service
@@ -77,7 +74,6 @@ webhooks:
   sideEffects: NoneOnDryRun
 - admissionReviewVersions:
   - v1
-  - v1beta1
   clientConfig:
     service:
       name: webhook-service
