+++ k8s/controllers/k8s/utils.go
@@ -1,4 +1,4 @@
-package controllers
+package k8s
 
 import (
 	"crypto/rand"
